+++ src/tests/test_suite.ts
@@ -16,33 +16,51 @@ export interface TestConfig {
 }
 
 export interface TestOptions {
-	timeout?: number,
+	timeout?: Milliseconds,
 	config?: TestConfig,
 	notifyFn?: NotifyFn,
 	clock?: Clock,
 }
 
+export interface DescribeOptions {
+	timeout?: number,
+}
+
+export interface ItOptions {
+	timeout?: number,
+}
+
 export type NotifyFn = (testResult: TestCaseResult) => void;
 
 interface Describe {
-	(optionalName?: string | DescribeFn, describeFn?: DescribeFn): TestSuite,
-	skip: (optionalName?: string | DescribeFn, descrbeFn?: DescribeFn) => TestSuite,
-	only: (optionalName?: string | DescribeFn, describeFn?: DescribeFn) => TestSuite,
+	(
+		optionalName?: string | DescribeOptions | DescribeFn,
+		optionalOptions?: DescribeOptions | DescribeFn,
+		describeFn?: DescribeFn
+	): TestSuite,
+	skip: (
+		optionalName?: string | DescribeOptions | DescribeFn,
+		optionalOptions?: DescribeOptions | DescribeFn,
+		descrbeFn?: DescribeFn
+	) => TestSuite,
+	only: (
+		optionalName?: string | DescribeOptions | DescribeFn,
+		optionalOptions?: DescribeOptions | DescribeFn,
+		describeFn?: DescribeFn
+	) => TestSuite,
 }
-type DescribeFn = (suiteUtilities: SuiteParameters) => void;
+type DescribeFn = () => void;
 
 interface It {
-	(name: string, itFn?: ItFn): void;
-	skip: (name: string, itFn?: ItFn) => void,
-	only: (name: string, itFn?: ItFn) => void,
+	(name: string, optionalOptions?: ItOptions | ItFn, itFn?: ItFn): void;
+	skip: (name: string, optionalOptions?: ItOptions | ItFn, itFn?: ItFn) => void,
+	only: (name: string, optionalOptions?: ItOptions | ItFn, itFn?: ItFn) => void,
 }
 type ItFn = (testUtilities: TestParameters) => Promise<void> | void;
 
-type BeforeAfter = (fn: ItFn) => void;
+type BeforeAfter = (optionalOptions: ItOptions | ItFn, fnAsync?: ItFn) => void;
 
-interface SuiteParameters {
-	setTimeout: (newTimeout: Milliseconds) => void,
-}
+type BeforeAfterDefinition = { options: ItOptions, fnAsync: ItFn };
 
 interface TestParameters {
 	getConfig: <T>(key: string) => T,
@@ -62,8 +80,8 @@ interface RecursiveRunOptions {
 interface Runnable {
 	_recursiveRunAsync: (
 		parentMark: TestMarkValue,
-		parentBeforeEachFns: ItFn[],
-		parentAfterEachFns: ItFn[],
+		parentBeforeEachFns: BeforeAfterDefinition[],
+		parentAfterEachFns: BeforeAfterDefinition[],
 		options: RecursiveRunOptions,
 	) => Promise<TestResult> | TestResult;
 	_isDotOnly: () => boolean,
@@ -91,10 +109,10 @@ export class TestSuite implements Runnable {
 	private _tests: Runnable[];
 	private _hasDotOnlyChildren: boolean;
 	private _allChildrenSkipped: boolean;
-	private _beforeAllFns: ItFn[];
-	private _afterAllFns: ItFn[];
-	private _beforeEachFns: ItFn[];
-	private _afterEachFns: ItFn[];
+	private _beforeAllFns: BeforeAfterDefinition[];
+	private _afterAllFns: BeforeAfterDefinition[];
+	private _beforeEachFns: BeforeAfterDefinition[];
+	private _afterEachFns: BeforeAfterDefinition[];
 	private _timeout?: Milliseconds;
 	private _filename?: string;
 
@@ -148,26 +166,23 @@ export class TestSuite implements Runnable {
 
 	/** @private */
 	static _create(
-		nameOrSuiteFn: string | DescribeFn | undefined,
+		nameOrOptionsOrSuiteFn: string | DescribeOptions | DescribeFn | undefined,
+		optionsOrSuiteFn: DescribeOptions | DescribeFn | undefined,
 		possibleSuiteFn: DescribeFn | undefined,
 		mark: TestMarkValue,
 	): TestSuite {
-		ensure.signature(arguments, [ [ undefined, String, Function ], [ undefined, Function ], String ]);
+		const DescribeOptionsType = { timeout: Number };
+		ensure.signature(arguments, [
+			[ undefined, DescribeOptionsType, String, Function ],
+			[ undefined, DescribeOptionsType, Function ],
+			[ undefined, Function ],
+			String
+		]);
 
-		let name: string;
-		let suiteFn: DescribeFn | undefined;
-
-		if (nameOrSuiteFn instanceof Function || (nameOrSuiteFn === undefined && possibleSuiteFn === undefined)) {
-			name = "";
-			suiteFn = nameOrSuiteFn;
-		}
-		else {
-			name = nameOrSuiteFn ?? "";
-			suiteFn = possibleSuiteFn;
-		}
+		const { name, options, suiteFn } = decipherOverloadedParameters();
 
 		if (suiteFn !== undefined) {
-			return this.#runDescribeFunction(suiteFn, name, mark);
+			return this.#runDescribeFunction(suiteFn, name, mark, options.timeout);
 		}
 		else if (mark === TestMark.only) {
 			return new TestSuite(name, mark, {
@@ -175,7 +190,52 @@ export class TestSuite implements Runnable {
 			});
 		}
 		else {
-			return new TestSuite(name, TestMark.skip, {});
+			return new TestSuite(name, TestMark.skip, { timeout: options.timeout });
+		}
+
+		function decipherOverloadedParameters() {
+			let name: string;
+			let options: DescribeOptions | undefined;
+			let suiteFn: DescribeFn | undefined;
+
+			switch (typeof nameOrOptionsOrSuiteFn) {
+				case "string":
+					name = nameOrOptionsOrSuiteFn;
+					break;
+				case "object":
+					options = nameOrOptionsOrSuiteFn;
+					break;
+				case "function":
+					suiteFn = nameOrOptionsOrSuiteFn;
+					break;
+				case "undefined":
+					break;
+				default:
+					ensure.unreachable(`Unknown typeof for nameOrOptionsOrSuiteFn: ${typeof nameOrOptionsOrSuiteFn}`);
+			}
+			switch (typeof optionsOrSuiteFn) {
+				case "object":
+					ensure.that(options === undefined, "Received two options parameters");
+					options = optionsOrSuiteFn;
+					break;
+				case "function":
+					ensure.that(suiteFn === undefined, "Received two suite function parameters");
+					suiteFn = optionsOrSuiteFn;
+					break;
+				case "undefined":
+					break;
+				default:
+					ensure.unreachable(`Unknown typeof for optionsOrSuiteFn: ${typeof optionsOrSuiteFn}`);
+			}
+			if (possibleSuiteFn !== undefined) {
+				ensure.that(suiteFn === undefined, "Received two suite function parameters");
+				suiteFn = possibleSuiteFn;
+			}
+
+			name ??= "";
+			options ??= {};
+
+			return { name, options, suiteFn };
 		}
 	}
 
@@ -183,49 +243,70 @@ export class TestSuite implements Runnable {
 		describeFn: DescribeFn,
 		name: string,
 		mark: TestMarkValue,
+		timeout?: Milliseconds,
 	): TestSuite {
 		const tests: Runnable[] = [];
-		const beforeAllFns: ItFn[] = [];
-		const afterAllFns: ItFn[] = [];
-		const beforeEachFns: ItFn[] = [];
-		const afterEachFns: ItFn[] = [];
-		let timeout: number | undefined;
+		const beforeAllFns: BeforeAfterDefinition[] = [];
+		const afterAllFns: BeforeAfterDefinition[] = [];
+		const beforeEachFns: BeforeAfterDefinition[] = [];
+		const afterEachFns: BeforeAfterDefinition[] = [];
 
 		const pushTest = <T extends Runnable>(test: T): T => {
 			tests.push(test);
 			return test;
 		};
 
-		const result: Describe = (optionalName, suiteFn) => this._create(optionalName, suiteFn, TestMark.none);
-		result.skip = (optionalName, suiteFn) => this._create(optionalName, suiteFn, TestMark.skip);
-		result.only = (optionalName, suiteFn) => this._create(optionalName, suiteFn, TestMark.only);
+		const result: Describe = (optionalName, optionalOptions, fn) => this._create(optionalName, optionalOptions, fn, TestMark.none);
+		result.skip = (optionalName, optionalOptions, fn) => this._create(optionalName, optionalOptions, fn, TestMark.skip);
+		result.only = (optionalName, optionalOptions, fn) => this._create(optionalName, optionalOptions, fn, TestMark.only);
 
-		const describe: Describe = (optionalName, suiteFn) => pushTest(TestSuite._create(optionalName, suiteFn, TestMark.none));
-		describe.skip = (optionalName, describeFn) => pushTest(TestSuite._create(optionalName, describeFn, TestMark.skip));
-		describe.only = (optionalName, suiteFn) => pushTest(TestSuite._create(optionalName, suiteFn, TestMark.only));
+		const describe: Describe = (optionalName, optionalOptions, fn) => pushTest(TestSuite._create(optionalName, optionalOptions, fn, TestMark.none));
+		describe.skip = (optionalName, optionalOptions, fn) => pushTest(TestSuite._create(optionalName, optionalOptions, fn, TestMark.skip));
+		describe.only = (optionalName, optionalOptions, fn) => pushTest(TestSuite._create(optionalName, optionalOptions, fn, TestMark.only));
 
-		const it: It = (name, testCaseFn) => pushTest(new TestCase(name, testCaseFn, TestMark.none));
-		it.skip = (name, testCaseFn) => pushTest(new TestCase(name, testCaseFn, TestMark.skip));
-		it.only = (name, testCaseFn) => pushTest(new TestCase(name, testCaseFn, TestMark.only));
+		const it: It = (name, optionalOptions, testCaseFn) => pushTest(new TestCase(name, optionalOptions, testCaseFn, TestMark.none));
+		it.skip = (name, optionalOptions, testCaseFn) => pushTest(new TestCase(name, optionalOptions, testCaseFn, TestMark.skip));
+		it.only = (name, optionalOptions, testCaseFn) => pushTest(new TestCase(name, optionalOptions, testCaseFn, TestMark.only));
 
 		testContext.push({
 			describe,
 			it,
-			beforeAll: (fnAsync) => { beforeAllFns.push(fnAsync); },
-			afterAll: (fnAsync) => { afterAllFns.push(fnAsync); },
-			beforeEach: (fnAsync) => { beforeEachFns.push(fnAsync); },
-			afterEach: (fnAsync) => { afterEachFns.push(fnAsync); },
+			beforeAll: defineBeforeAfterFn(beforeAllFns),
+			afterAll: defineBeforeAfterFn(afterAllFns),
+			beforeEach: defineBeforeAfterFn(beforeEachFns),
+			afterEach: defineBeforeAfterFn(afterEachFns),
 		});
 		try {
-			describeFn({
-				setTimeout: (newTimeoutInMs) => { timeout = newTimeoutInMs; },
-			});
+			describeFn();
 		}
 		finally {
 			testContext.pop();
 		}
 
 		return new TestSuite(name, mark, { tests, beforeAllFns, afterAllFns, beforeEachFns, afterEachFns, timeout });
+
+		function defineBeforeAfterFn(beforeAfterArray: BeforeAfterDefinition[]) {
+			return function (optionsOrFnAsync: ItOptions | ItFn, possibleFnAsync?: ItFn) {
+				ensure.signature(arguments, [
+					[ { timeout: Number }, Function ],
+					[ undefined, Function ],
+				]);
+
+				let options: ItOptions;
+				let fnAsync: ItFn;
+
+				if (possibleFnAsync === undefined) {
+					options = {};
+					fnAsync = optionsOrFnAsync as ItFn;
+				}
+				else {
+					options = optionsOrFnAsync as ItOptions;
+					fnAsync = possibleFnAsync;
+				}
+
+				beforeAfterArray.push({ options, fnAsync });
+			};
+		}
 	}
 
 	/** Internal use only. (Use {@link TestSuite.create} or {@link TestSuite.fromModulesAsync} instead.) */
@@ -238,13 +319,12 @@ export class TestSuite implements Runnable {
 		timeout,
 	}: {
 		tests?: Runnable[],
-		beforeAllFns?: ItFn[],
-		afterAllFns?: ItFn[],
-		beforeEachFns?: ItFn[],
-		afterEachFns?: ItFn[],
+		beforeAllFns?: BeforeAfterDefinition[],
+		afterAllFns?: BeforeAfterDefinition[],
+		beforeEachFns?: BeforeAfterDefinition[],
+		afterEachFns?: BeforeAfterDefinition[],
 		timeout?: Milliseconds,
 	}) {
-
 		this._name = name;
 		this._mark = mark;
 		this._tests = tests;
@@ -305,8 +385,8 @@ export class TestSuite implements Runnable {
 	/** @private */
 	async _recursiveRunAsync(
 		parentMark: TestMarkValue,
-		parentBeforeEachFns: ItFn[],
-		parentAfterEachFns: ItFn[],
+		parentBeforeEachFns: BeforeAfterDefinition[],
+		parentAfterEachFns: BeforeAfterDefinition[],
 		options: RecursiveRunOptions,
 	) {
 		const name = [ ...options.name ];
@@ -350,14 +430,42 @@ export class TestSuite implements Runnable {
 class TestCase implements Runnable {
 
 	protected _name: string;
+	private _timeout?: Milliseconds;
 	private _testFn?: ItFn;
 	private _mark: TestMarkValue;
 
-	constructor(name: string, testFn: ItFn | undefined, mark: TestMarkValue) {
-		ensure.signature(arguments, [ String, [ undefined, Function ], String ]);
+	constructor(
+		name: string,
+		optionsOrTestFn: TestOptions | ItFn | undefined,
+		possibleTestFn: ItFn | undefined,
+		mark: TestMarkValue
+	) {
+		ensure.signature(arguments, [
+			String,
+			[ undefined, { timeout: [ undefined, Number ]}, Function ],
+			[ undefined, Function ],
+			String
+		]);
 
 		this._name = name;
-		this._testFn = testFn;
+
+		switch (typeof optionsOrTestFn) {
+			case "object":
+				this._timeout = optionsOrTestFn.timeout;
+				break;
+			case "function":
+				this._testFn = optionsOrTestFn;
+				break;
+			case "undefined":
+				break;
+			default:
+				ensure.unreachable(`Unknown typeof optionsOrTestFn: ${typeof optionsOrTestFn}`);
+		}
+		if (possibleTestFn !== undefined) {
+			ensure.that(this._testFn === undefined, "Received two test function parameters");
+			this._testFn = possibleTestFn;
+		}
+
 		this._mark = mark;
 	}
 
@@ -376,8 +484,8 @@ class TestCase implements Runnable {
 	/** @private */
 	async _recursiveRunAsync(
 		parentMark: TestMarkValue,
-		beforeEachFns: ItFn[],
-		afterEachFns: ItFn[],
+		beforeEachFns: BeforeAfterDefinition[],
+		afterEachFns: BeforeAfterDefinition[],
 		options: RecursiveRunOptions,
 	): Promise<TestCaseResult> {
 		const name = [ ...options.name ];
@@ -409,7 +517,7 @@ class TestCase implements Runnable {
 			const beforeResult = await runBeforeOrAfterFnsAsync(options.name, beforeEachFns, self._mark, options);
 			if (!isSuccess(beforeResult)) return beforeResult;
 
-			const itResult = await runTestFnAsync(options.name, self._testFn!, self._mark, options);
+			const itResult = await runTestFnAsync(options.name, self._testFn!, self._mark, self._timeout, options);
 			const afterResult = await runBeforeOrAfterFnsAsync(options.name, afterEachFns, self._mark, options);
 
 			if (!isSuccess(itResult)) return itResult;
@@ -425,7 +533,7 @@ class FailureTestCase extends TestCase {
 	private _error: unknown;
 
 	constructor(name: string, error: unknown, filename?: string) {
-		super(name, undefined, TestMark.none);
+		super(name, undefined, undefined, TestMark.none);
 
 		this._filename = filename;
 		this._error = error;
@@ -433,8 +541,8 @@ class FailureTestCase extends TestCase {
 
 	override async _recursiveRunAsync(
 		parentMark: TestMarkValue,
-		beforeEachFns: ItFn[],
-		afterEachFns: ItFn[],
+		beforeEachFns: BeforeAfterDefinition[],
+		afterEachFns: BeforeAfterDefinition[],
 		options: RecursiveRunOptions,
 	): Promise<TestCaseResult> {
 		const result = TestResult.fail([ this._name ], this._error, this._filename);
@@ -447,12 +555,12 @@ class FailureTestCase extends TestCase {
 
 async function runBeforeOrAfterFnsAsync(
 	name: string[],
-	fns: ItFn[],
+	beforeAfterArray: BeforeAfterDefinition[],
 	mark: TestMarkValue,
 	options: RecursiveRunOptions,
 ): Promise<TestCaseResult> {
-	for await (const fn of fns) {
-		const result = await runTestFnAsync(name, fn, mark, options);
+	for await (const beforeAfter of beforeAfterArray) {
+		const result = await runTestFnAsync(name, beforeAfter.fnAsync, mark, beforeAfter.options.timeout, options);
 		if (!isSuccess(result)) return result;
 	}
 	return TestResult.pass(name, options.filename, mark);
@@ -462,12 +570,15 @@ async function runTestFnAsync(
 	name: string[],
 	fn: ItFn,
 	mark: TestMarkValue,
+	testTimeout: Milliseconds | undefined,
 	{ clock, filename, timeout, config }: RecursiveRunOptions,
 ): Promise<TestCaseResult> {
 	const getConfig = <T>(name: string) => {
 		if (config[name] === undefined) throw new Error(`No test config found for name '${name}'`);
 		return config[name] as T;
 	};
+
+	timeout = testTimeout ?? timeout;
 
 	return await clock.timeoutAsync(timeout, async () => {
 		try {
@@ -488,14 +599,15 @@ function isSuccess(result: TestCaseResult) {
 
 
 function startTest(
-	nameOrSuiteFn: string | DescribeFn | undefined,
-	possibleSuiteFn: DescribeFn | undefined,
+	optionalName: string | DescribeOptions | DescribeFn | undefined,
+	optionalOptions: DescribeOptions | DescribeFn | undefined,
+	fn: DescribeFn | undefined,
 	mark: TestMarkValue,
 ): TestSuite {
 	ensure.that(testContext.length === 0, "test() is not re-entrant [don't run test() inside of test()]");
 
 	try {
-		return TestSuite._create(nameOrSuiteFn, possibleSuiteFn, mark);
+		return TestSuite._create(optionalName, optionalOptions, fn, mark);
 	}
 	finally {
 		ensure.that(testContext.length === 0, "test() didn't clear its context; must be an error in ergotest");
@@ -505,99 +617,138 @@ function startTest(
 /**
  * Creates a top-level test suite. In your test module, call this function and `export default` the result. Add `.skip`
  * to skip this test suite and `.only` to only run this test suite.
- * @param {string} [optionalName] The name of the test suite. You can skip this parameter and pass {@link fn} instead.
- * @param {function} [fn] The body of the test suite. In the body, call {@link describe}, {@link it}, {@link beforeAll},
- *   {@link afterAll}, {@link beforeEach}, and {@link afterEach} to define the tests in the suite. If undefined, this
- *   test suite will be skipped.
+ * @param {string} [optionalName] The name of the test suite. You can skip this parameter and pass
+ *   {@link optionalOptions} or {@link fn} instead.
+ * @param {DescribeOptions} [optionalOptions] The test suite options. You can skip this parameter and pass {@link fn}
+ *   instead.
+ * @param {function} [fn] The body of the test suite. In the body, call {@link describe}, {@link it}, {@link
+ *   beforeAll}, {@link afterAll}, {@link beforeEach}, and {@link afterEach} to define the tests in the suite. If
+ *   undefined, this test suite will be skipped.
  * @returns {TestSuite} The test suite. You’ll typically `export default` the return value rather than using it
  *   directly.
  */
-export function test(optionalName?: string | DescribeFn, fn?: DescribeFn) {
-	return startTest(optionalName, fn, TestMark.none);
+export function test(
+	optionalName?: string | DescribeOptions | DescribeFn,
+	optionalOptions?: DescribeOptions | DescribeFn,
+	fn?: DescribeFn,
+): TestSuite {
+	return startTest(optionalName, optionalOptions, fn, TestMark.none);
 }
 
-test.skip = function(optionalName?: string | DescribeFn, fn?: DescribeFn) {
-	return startTest(optionalName, fn, TestMark.skip);
+test.skip = function(
+	optionalName?: string | DescribeOptions | DescribeFn,
+	optionalOptions?: DescribeOptions | DescribeFn,
+	fn?: DescribeFn,
+): TestSuite {
+	return startTest(optionalName, optionalOptions, fn, TestMark.skip);
 };
 
-test.only = function(optionalName?: string | DescribeFn, fn?: DescribeFn) {
-	return startTest(optionalName, fn, TestMark.only);
+test.only = function(
+	optionalName?: string | DescribeOptions | DescribeFn,
+	optionalOptions?: DescribeOptions | DescribeFn,
+	fn?: DescribeFn,
+): TestSuite {
+	return startTest(optionalName, optionalOptions, fn, TestMark.only);
 };
 
 /**
  * Adds a nested test suite to the current test suite. Must be run inside of a {@link test} or {@link describe}
  * function. Add `.skip` to skip this test suite and `.only` to only run this test suite.
- * @param {string} [optionalName] The name of the test suite. You can skip this parameter and pass {@link fn} instead.
- * @param {function} [fn] The body of the test suite. In the body, call {@link describe}, {@link it}, {@link beforeAll},
- *   {@link afterAll}, {@link beforeEach}, and {@link afterEach} to define the tests in the suite. If undefined, this
- *   test suite will be skipped.
+ * @param {string} [optionalName] The name of the test suite. You can skip this parameter and pass
+ *   {@link optionalOptions} or {@link fn} instead.
+ * @param {DescribeOptions} [optionalOptions] The test suite options. You can skip this parameter and pass {@link fn}
+ *   instead.
+ * @param {function} [fn] The body of the test suite. In the body, call {@link describe}, {@link it}, {@link
+ *   beforeAll}, {@link afterAll}, {@link beforeEach}, and {@link afterEach} to define the tests in the suite. If
+ *   undefined, this test suite will be skipped.
  * @returns {TestSuite} The test suite. You’ll typically ignore the return value.
  */
-export function describe(optionalName?: string | DescribeFn, fn?: DescribeFn) {
-	currentContext("describe").describe(optionalName, fn);
+export function describe(
+	optionalName?: string | DescribeOptions | DescribeFn,
+	optionalOptions?: DescribeOptions | DescribeFn,
+	fn?: DescribeFn,
+) {
+	currentContext("describe").describe(optionalName, optionalOptions, fn);
 }
 
-describe.skip = function(optionalName?: string | DescribeFn, fn?: DescribeFn) {
-	currentContext("describe").describe.skip(optionalName, fn);
+describe.skip = function(
+	optionalName?: string | DescribeOptions | DescribeFn,
+	optionalOptions?: DescribeOptions | DescribeFn,
+	fn?: DescribeFn,
+) {
+	currentContext("describe").describe.skip(optionalName, optionalOptions, fn);
 };
 
-describe.only = function(optionalName?: string | DescribeFn, fn?: DescribeFn) {
-	currentContext("describe").describe.only(optionalName, fn);
+describe.only = function(
+	optionalName?: string | DescribeOptions | DescribeFn,
+	optionalOptions?: DescribeOptions | DescribeFn,
+	fn?: DescribeFn,
+) {
+	currentContext("describe").describe.only(optionalName, optionalOptions, fn);
 };
 
 /**
  * Adds a test to the current test suite. Must be run inside of a {@link test} or {@link describe} function. Add
  * `.skip` to skip this test and `.only` to only run this test.
  * @param {string} name The name of the test.
+ * @param {ItOptions} [optionalOptions] The test options. You can skip this parameter and pass {@link fnAsync} instead.
  * @param {function} [fnAsync] The body of the test. May be synchronous or asynchronous. If undefined, this test will be
  *   skipped.
  */
-export function it(name: string, fnAsync?: ItFn) {
-	currentContext("it").it(name, fnAsync);
+export function it(name: string, optionalOptions?: ItOptions | ItFn, fnAsync?: ItFn) {
+	currentContext("it").it(name, optionalOptions, fnAsync);
 }
 
-it.skip = function(name: string, fnAsync?: ItFn) {
-	currentContext("it").it.skip(name, fnAsync);
+it.skip = function it(name: string, optionalOptions?: ItOptions | ItFn, fnAsync?: ItFn) {
+	currentContext("it").it.skip(name, optionalOptions, fnAsync);
 };
 
-it.only = function(name: string, fnAsync?: ItFn) {
-	currentContext("it").it.only(name, fnAsync);
+it.only = function it(name: string, optionalOptions?: ItOptions | ItFn, fnAsync?: ItFn) {
+	currentContext("it").it.only(name, optionalOptions, fnAsync);
 };
 
 /**
  * Adds a function to run before all the tests in the current test suite. Must be run inside of a {@link test} or
  * {@link describe} function.
- * @param {function} [fnAsync] The function to run. May be synchronous or asynchronous.
+ * @param {ItOptions} [optionalOptions] The before/after options. You can skip this parameter and pass @{link fnAsync}
+ *   instead.
+ * @param {function} fnAsync The function to run. May be synchronous or asynchronous.
  */
-export function beforeAll(fnAsync: ItFn) {
-	currentContext("beforeAll").beforeAll(fnAsync);
+export function beforeAll(optionalOptions: ItOptions | ItFn, fnAsync?: ItFn) {
+	currentContext("beforeAll").beforeAll(optionalOptions, fnAsync);
 }
 
 /**
  * Adds a function to run after all the tests in the current test suite. Must be run inside of a {@link test} or
  * {@link describe} function.
+ * @param {ItOptions} [optionalOptions] The before/after options. You can skip this parameter and pass @{link fnAsync}
+ *   instead.
  * @param {function} [fnAsync] The function to run. May be synchronous or asynchronous.
  */
-export function afterAll(fnAsync: ItFn) {
-	currentContext("afterAll").afterAll(fnAsync);
+export function afterAll(optionalOptions: ItOptions | ItFn, fnAsync?: ItFn) {
+	currentContext("afterAll").afterAll(optionalOptions, fnAsync);
 }
 
 /**
  * Adds a function to run bfeore each of the tests in the current test suite. Must be run inside of a {@link test} or
  * {@link describe} function.
+ * @param {ItOptions} [optionalOptions] The before/after options. You can skip this parameter and pass @{link fnAsync}
+ *   instead.
  * @param {function} [fnAsync] The function to run. May be synchronous or asynchronous.
  */
-export function beforeEach(fnAsync: ItFn) {
-	currentContext("beforeEach").beforeEach(fnAsync);
+export function beforeEach(optionalOptions: ItOptions | ItFn, fnAsync?: ItFn) {
+	currentContext("beforeEach").beforeEach(optionalOptions, fnAsync);
 }
 
 /**
  * Adds a function to run after each of the tests in the current test suite. Must be run inside of a {@link test} or
  * {@link describe} function.
+ * @param {ItOptions} [optionalOptions] The before/after options. You can skip this parameter and pass @{link fnAsync}
+ *   instead.
  * @param {function} [fnAsync] The function to run. May be synchronous or asynchronous.
  */
-export function afterEach(fnAsync: ItFn) {
-	currentContext("afterEach").afterEach(fnAsync);
+export function afterEach(optionalOptions: ItOptions | ItFn, fnAsync?: ItFn) {
+	currentContext("afterEach").afterEach(optionalOptions, fnAsync);
 }
 
 function currentContext(functionName: string) {